Move UseTypeAdapter annotation to annotation package
